Produce more detailed output for API request errors
Unfortunately with the current HTTP API client
error type, it's impossible to deserialize
the response body: we are provided with a reference
to an Option<Response> and .unwrap() and such
consume the value.

Ideally the client should provide a payload
as part of the error, since it owns the response
and easily copy the payload.