Selectively support environment variables for key global options
This is preparatory work and not really a feature
in and of itself. Let me explain.

Environment variables have objective downsides:

1. Mostly non-existent type checking
2. Hard to inspect
3. Hard to reason about, given that shell environment
   inheritance is always an afterthought

Therefore using environment variables for CLI arguments
is a catastrophic disservice to the user most of
the time, and they should be used as little as
possible.

However, for "pre-flight" global arguments,
such as the --config-file default, they are a
reasonable or even the only alternative
available.

This commit enables Clap's environment variable
feature for a few global flags.

One reason to have this feature is that
in order to enable CLI parser-level features,
such as subcommand name inference or
--long-option inference, both of which Clap
supports and both of which can be a solid
time saver for heavy users, we need to
have a "pre-CLI args" and "pre-config file"
configuration mechanism, and environment
variables is the only option that's universally
known.

The end goal of this change is not so much
to allow people set RABBITMQADMIN_TARGET_HOST
but rather enable features such as

 * RABBITMQADMIN_NON_INTERACTIVE_MODE
 * RABBITMQADMIN_CONFIG_FILE_PATH
 * RABBITMQADMIN_INFER_CLI_SUBCOMMANDS
 * RABBITMQADMIN_INFER_CLI_LONG_ARGUMENTS

For --host, --port, --username and such,
use a config file and steer clear of environment
variables for the aforementioned reasons.