Make it possible to specify a few global options --at-the-end
This generation of 'rabbitmqadmin' strictly separates global options
from the command-specific ones. This has both pros and cons.

Specifying a few very common options, namely --vhost, at the end
of the argument list, was common with 'rabbitmqadmin' v1.

Clap allows a global option to be marked as such, which means it
will still be correctly parsed even when given amongst
the command-specific ones.

This has a serious downside, too: if a subcommand declares
an option with the same ID, long or short name, the parser
will panic.

Therefore this commit cherry picks a few options that are
truly global and unlikely will run into conflicts:

 * --vhost
 * TLS-related options
 * --quiet/-q

The docs will continue saying that all global options must be
passed before the subcommands because that's still true for
the majority of them.